+++ SonsSdk/SdkEvents.cs
@@ -8,11 +8,13 @@ using Sons.Events;
 using Sons.Gui.Options;
 using Sons.Inventory;
 using Sons.Loading;
+using Sons.Save;
 using TheForest.Utils;
 using TMPro;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.SceneManagement;
+using Action = Il2CppSystem.Action;
 using Color = UnityEngine.Color;
 using Int32 = Il2CppSystem.Int32;
 using Object = UnityEngine.Object;
@@ -60,6 +62,33 @@ public static class SdkEvents
     /// Called by HDRP at the end of rendering a frame
     /// </summary>
     public static readonly MelonEvent<ScriptableRenderContext, Il2CppSystem.Collections.Generic.List<Camera>> OnCameraRender = new();
+    
+    /// <summary>
+    /// Called before serializing the save game
+    /// </summary>
+    public static readonly MelonEvent BeforeSaveLoading = new();
+    
+    /// <summary>
+    /// Called after serializing the save game.
+    /// Returns if it should only save the player (i.e. saveGameType = MultiplayerClient)
+    /// </summary>
+    public static readonly MelonEvent<bool> AfterSaveLoading = new();
+    
+    /// <summary>
+    /// Called before deserializing the save game
+    /// </summary>
+    public static readonly MelonEvent BeforeLoadSave = new();
+    
+    /// <summary>
+    /// Called after deserializing the save game
+    /// </summary>
+    public static readonly MelonEvent AfterLoadSave = new();
+
+    /// <summary>
+    /// Called on the fith activation sequence when the game mode is initialized.
+    /// This should be used for general world modifications.
+    /// </summary>
+    public static readonly MelonEvent OnGameActivated = new();
 
     public static readonly MelonEvent<ESonsScene> OnSonsSceneInitialized = new();
 
@@ -80,7 +109,7 @@ public static class SdkEvents
         EventRegistry.Register(GameEvent.ArmorEquipped, (EventRegistry.SubscriberCallback)SonsEventOnArmorEquipped);
 
         RenderPipelineManager.endContextRendering += (Il2CppSystem.Action<ScriptableRenderContext, Il2CppSystem.Collections.Generic.List<Camera>>)OnEndContextRendering;
-        
+
         _isInitialized = true;
     }
 
@@ -96,9 +125,19 @@ public static class SdkEvents
                 OnSonsSceneInitialized.Invoke(ESonsScene.Loading);
                 break;
             case GameSceneName:
+                RLog.Msg(System.Drawing.Color.DeepPink, $"=== Game Scene Initialized ===");
+                LoadSave.OnGameStart += (Action)OnGameActivation;
                 OnSonsSceneInitialized.Invoke(ESonsScene.Game);
                 break;
         }
+    }
+
+    internal static void RegisterSaveGameManager()
+    {
+        SaveGameManager.RegisterBeforeLoadCallback((Action)OnBeforeLoadSave);
+        SaveGameManager.RegisterAfterLoadCallback((Action)OnAfterLoadSave);
+        SaveGameManager.RegisterBeforeSaveCallback((Action)OnBeforeSave);
+        SaveGameManager.RegisterAfterSaveCallback((Action<bool>)OnAfterSave);
     }
 
     private static void OnUpdateInternal()
@@ -133,13 +172,11 @@ public static class SdkEvents
         }
         
         OnItemConsumed.Invoke(item);
-        RLog.Msg($"Item consumed: {item.Data.Name}");
     }
     
     private static void SonsEventOnArmorEquipped(object _)
     {
         OnArmorEquipped.Invoke();
-        RLog.Msg($"Armor equipped");
     }
     
     private static void SonsEventOnItemCrafted(Il2CppSystem.Object o)
@@ -156,6 +193,35 @@ public static class SdkEvents
     
     private static void OnEndContextRendering(ScriptableRenderContext context, Il2CppSystem.Collections.Generic.List<Camera> cameras){
         OnCameraRender.Invoke(context, cameras);
+    }
+    
+    private static void OnBeforeSave()
+    {
+        RLog.Debug($"ON_BEFORE_SAVE_LOADING");
+        BeforeSaveLoading.Invoke();
+    }
+    
+    private static void OnAfterSave(bool savePlayerOnly)
+    {
+        RLog.Debug($"ON_AFTER_SAVE_LOADING");
+        AfterSaveLoading.Invoke(savePlayerOnly);
+    }
+    
+    private static void OnBeforeLoadSave()
+    {
+        RLog.Debug($"ON_BEFORE_LOAD_SAVE");
+        BeforeLoadSave.Invoke();
+    }
+    
+    private static void OnAfterLoadSave()
+    {
+        RLog.Debug($"ON_AFTER_LOAD_SAVE");
+        AfterLoadSave.Invoke();
+    }
+
+    private static void OnGameActivation()
+    {
+        RLog.Msg(System.Drawing.Color.DeepPink, $"||| Game Activated |||");
     }
 
     #endregion